// segfault-backend/src/data/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  GUEST: 'GUEST',
  PIGS: 'PIGS'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const IssueStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED'
} as const

export type IssueStatus = (typeof IssueStatus)[keyof typeof IssueStatus]


export const IssueAuthorized = {
  TRUE: 'TRUE',
  FALSE: 'FALSE'
} as const

export type IssueAuthorized = (typeof IssueAuthorized)[keyof typeof IssueAuthorized]


export const IssueError = {
  NONE: 'NONE',
  INVALID_LOCATION: 'INVALID_LOCATION',
  INAPPROPRIATE_CONTENT: 'INAPPROPRIATE_CONTENT',
  PENDING: 'PENDING'
} as const

export type IssueError = (typeof IssueError)[keyof typeof IssueError]


export const IssueType = {
  POTHOLE: 'POTHOLE',
  ROAD_DAMAGE: 'ROAD_DAMAGE',
  STREETLIGHT_FAULT: 'STREETLIGHT_FAULT',
  GARBAGE_UNCOLLECTED: 'GARBAGE_UNCOLLECTED',
  ILLEGAL_DUMPING: 'ILLEGAL_DUMPING',
  DRAINAGE_BLOCKED: 'DRAINAGE_BLOCKED',
  SEWAGE_OVERFLOW: 'SEWAGE_OVERFLOW',
  WATER_SUPPLY_ISSUE: 'WATER_SUPPLY_ISSUE',
  LOW_WATER_PRESSURE: 'LOW_WATER_PRESSURE',
  OPEN_MANHOLE: 'OPEN_MANHOLE',
  BROKEN_FOOTPATH: 'BROKEN_FOOTPATH',
  ILLEGAL_ENCROACHMENT: 'ILLEGAL_ENCROACHMENT',
  STRAY_CATTLE: 'STRAY_CATTLE',
  TREE_FALL: 'TREE_FALL',
  TRAFFIC_LIGHT_FAULT: 'TRAFFIC_LIGHT_FAULT',
  MOSQUITO_BREEDING: 'MOSQUITO_BREEDING',
  NOISE_COMPLAINT: 'NOISE_COMPLAINT',
  BUILDING_SAFETY: 'BUILDING_SAFETY'
} as const

export type IssueType = (typeof IssueType)[keyof typeof IssueType]


export const NotificationType = {
  ISSUE_STATUS_UPDATE: 'ISSUE_STATUS_UPDATE',
  NEW_COMMENT: 'NEW_COMMENT',
  GENERAL: 'GENERAL',
  BAN_NOTICE: 'BAN_NOTICE',
  REMOVAL_NOTICE: 'REMOVAL_NOTICE',
  UPVOTE_RECEIVED: 'UPVOTE_RECEIVED'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]